feat: Implement backend for order creation
This commit introduces the backend functionality for creating orders.

- Adds `Order`, `OrderStatus`, `OrderAddress`, `AddressType`, `OrderItem`, and `Ticket` entities.
- Creates a database migration to update the schema with the new tables.
- Implements `CreateOrderRequestDto` and `OrderConfirmationResponseDto` for API communication.
- Adds a new `OrdersController` with a secure `POST /api/orders` endpoint for creating orders.
- The endpoint calculates the final price securely on the server based on `PricingFactors`.
- Adds a `ProductsController` to provide `PackagingProducts` to the frontend.

+++ BarbariBahar.API/Controllers/UsersController.cs
@@ -25,7 +25,7 @@ namespace BarbariBahar.API.Controllers
         {
             // مرحله ۱: پیدا کردن ID کاربر از توکن (Claim)
             var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
-            if (userIdClaim == null || !int.TryParse(userIdClaim.Value, out int userId))
+            if (userIdClaim == null || !long.TryParse(userIdClaim.Value, out long userId))
             {
                 return Unauthorized(new { Message = "توکن نامعتبر یا فاقد ID کاربر است." });
             }
